refactor: remove annual log map
the map was only used to map dates to a bool, so it's just a plain set.
and with the move to c++23, we have acces to std::chrono::month_day.

std::set<std::chrono::month_day> does the job just as good if not better
but with less code to maintain.

Besides that, the 'date map' variables and members are renamed to just 'dates'
as they are not maping a date to anything, just containing a list of dates.